Ininitial commit

+++ src/main/java/com/matthew/bookdb/controller/BookController.java
@@ -1,26 +1,28 @@
 package com.matthew.bookdb.controller;
 
-import com.matthew.bookdb.model.Book;
+import com.matthew.bookdb.mysql.model.Book;
 import com.matthew.bookdb.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
 
 @RestController
-@RequestMapping("/api/books")
+@RequestMapping("/api/v1/")
 public class BookController {
 
     @Autowired
     private BookRepository bookRepository;
 
-    @GetMapping
+    @GetMapping("/books")
     public List<Book> getAllBooks() {
         return bookRepository.findAll();
     }
 
-    @GetMapping("/{id}")
+    @GetMapping("books/{id}")
     public Optional<Book> getBookById(@PathVariable Integer id) {
         return bookRepository.findById(id);
     }
@@ -30,7 +32,7 @@ public class BookController {
         return bookRepository.save(book);
     }
 
-    @PutMapping("/{id}")
+    @PutMapping("books/{id}")
     public Book updateBook(@PathVariable Integer id, @RequestBody Book bookDetails) {
         return bookRepository.findById(id).map(book -> {
             book.setTitle(bookDetails.getTitle());
@@ -41,8 +43,13 @@ public class BookController {
         }).orElseThrow(() -> new RuntimeException("Book not with id " + id));
     }
 
-    @DeleteMapping("/{id}")
-    public void deleteBook(@PathVariable Integer id) {
-        bookRepository.deleteById(id);
+    @DeleteMapping("books/{id}")
+    public ResponseEntity<String> deleteBook(@PathVariable Integer id) {
+        if (bookRepository.existsById(id)) {
+            bookRepository.deleteById(id);
+            return ResponseEntity.ok("Book with id " + id + " deleted successfully.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Book with id " + id + " not found.");
+        }
     }
 }
